src/master-data/staging/requirements.txt
requests>=2.31.0
tqdm>=4.66.0
python-dateutil>=2.9.0
psycopg2>=2.9.11